services/notification-service/requirements.txt
Flask==3.1.3
Flask-SocketIO==5.4.0
Flask-CORS==5.0.0
python-dotenv==1.0.1
redis==5.2.1
sentry-sdk[flask]==2.20.0
eventlet==0.36.1
gunicorn==22.0.0
